Skip reporting individual game data to start.gg
The mutation currently doesn't properly mark the start.gg set as
completed and responds with the error message “"Match data is out of
date. Please refresh the page and try again” despite modifying state on
start.gg. This change might be a workaround for that, but it raises the
question of how to correctly report best-of-3 matches.